requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pandas==2.2.2
networkx==3.3
scikit-learn==1.5.1
xgboost==2.1.0
numpy==1.26.4
qiskit>=2.3.0
qiskit-aer>=0.17.0
qiskit-algorithms==0.3.0
qiskit_optimization==0.6.1
qiskit-ibm-runtime>=0.29.0
matplotlib==3.9.0
scipy==1.14.0
jinja2==3.1.4
aiofiles==24.1.0
httpx==0.27.0
pylatexenc==2.10
groq>=1.0.0
python-dotenv>=1.0.0
a2wsgi>=1.10.0          # ASGI→WSGI bridge for PythonAnywhere
